backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
gunicorn==21.2.0
